+++ test/unit/pulled_search/config_override.py
@@ -24,12 +24,10 @@ else:
     import unittest
 
 # Third-party
-import mock
 
 # Local
 sys.path.append(os.getcwd())
 import pulled_search
-import lib.gen_libs as gen_libs
 import version
 
 __version__ = version.__version__
@@ -85,15 +83,18 @@ class UnitTest(unittest.TestCase):
                 self.monitor_dir = "/dir_path/monitor_dir"
 
         self.cfg = CfgTest()
-        self.args_array = {"-t": "name@domain"}
-        self.args_array2 = {"-t": "name@domain", "-m": "/new_path/doc_dir"}
-        self.args_array3 = {"-t": "name@domain", "-n": "/new_path/monitor_dir"}
-        self.args_array4 = {"-t": "name@domain", "-n": "/new_path/monitor_dir",
-                            "-m": "/new_path/doc_dir"}
+        self.name = "name@domain"
+        self.docdir = "/new_path/doc_dir"
+        self.monitordir = "/new_path/monitor_dir"
+        self.args_array = {"-t": self.name}
+        self.args_array2 = {"-t": self.name, "-m": self.docdir}
+        self.args_array3 = {"-t": self.name, "-n": self.monitordir}
+        self.args_array4 = {"-t": self.name, "-n": self.monitordir,
+                            "-m": self.docdir}
         self.results = "/dir_path/doc_dir"
         self.results2 = "/dir_path/monitor_dir"
-        self.results3 = "/new_path/doc_dir"
-        self.results4 = "/new_path/monitor_dir"
+        self.results3 = self.docdir
+        self.results4 = self.monitordir
 
     def test_all_changes(self):